docs(model): add Javadoc comments for Job, Task, RetryPolicy, Schedule, JobStatus, and TaskStatus
