--- src/main/java/com/distributedscheduler/model/JobStatus.java
@@ -1,5 +1,0 @@
-package main.java.com.distributedscheduler.model;
-
-public enum JobStatus {
-    PENDING, RUNNING, COMPLETED, FAILED
-}
