--- src/main/java/com/distributedscheduler/model/RetryPolicy.java
@@ -1,8 +1,0 @@
-package main.java.com.distributedscheduler.model;
-
-public class RetryPolicy {
-    private int maxRetries;
-    private int delaySeconds;
-
-    // Constructors, Getters, Setters
-}
