--- src/main/java/com/distributedscheduler/model/Schedule.java
@@ -1,9 +1,0 @@
-package main.java.com.distributedscheduler.model;
-
-public class Schedule {
-    private String cronExpression;
-    private boolean isRecurring;
-
-    // Constructors, Getters, Setters
-}
-
