--- src/main/java/com/distributedscheduler/model/TaskStatus.java
@@ -1,5 +1,0 @@
-package main.java.com.distributedscheduler.model;
-
-public enum TaskStatus {
-    PENDING, RUNNING, COMPLETED, FAILED, SKIPPED
-}
